refactor(assets): consolidate all runtime assets under src/assets/
Moves all embedded runtime assets from scattered locations across src/
into a single src/assets/ root with predictable subdirectories. The build
step now mirrors src/assets/ → dist/assets/ exactly instead of using a
broad catch-all glob over src/**/*.

New layout:
  src/assets/backends/   ← schtasks-template.xml, launchd-template.xml
  src/assets/help/       ← CLI extended help markdown files
  src/assets/hints/      ← cli-hints-full.md, cli-hints-short.md
  src/assets/profiles/   ← improve profile JSON files (default, quick, etc.)
  src/assets/prompts/    ← LLM prompt templates (extract-session, graph-extract)
  src/assets/tasks/      ← task template YAML files (graph-refresh-weekly)
  src/assets/wiki/       ← wiki template markdown files
  src/assets/workflows/  ← workflow-template.md

copy-assets.ts: glob changed from src/**/*.{md,xml,json,yml,yaml} to
src/assets/**/* with dest transform src/assets/ → dist/assets/.
To add a new embedded asset: put it in the relevant src/assets/ subfolder,
update the import path — no glob changes needed.

Import paths updated in: improve-profiles.ts, extract-prompt.ts,
schtasks.ts, launchd.ts, wiki-templates.ts, wiki.ts, authoring.ts,
cli-hints.ts, graph-extract.ts
