+++ scripts/copy-assets.ts
@@ -1,11 +1,11 @@
 #!/usr/bin/env bun
 // Build-time asset step:
-//   1. Copy non-TS asset files from src/ to dist/ after tsc. Includes:
-//      - *.md, *.xml  — documentation and Windows task XML templates
-//      - *.json       — embedded profile files (src/commands/profiles/)
-//      - *.yml/*.yaml — embedded task templates (src/tasks/templates/)
-//      Prefer small embedded files over large in-source string constants;
-//      keep the file-system import pattern consistent across the project.
+//   1. Mirror src/assets/ → dist/assets/ after tsc.
+//      All runtime assets (profiles, task templates, backend templates,
+//      prompts, hints, wiki templates) live under src/assets/ with
+//      predictable subfolders. Output is always dist/assets/<subfolder>/<file>.
+//      To add a new embedded asset: put it in src/assets/, update the
+//      importing .ts file's path, done — no glob changes needed.
 //   2. Bundle scripts/migrate-storage.ts + scripts/migrations/*.ts into
 //      dist/scripts/ so globally-installed users (npm / prebuilt binary)
 //      can run them without `../src/...` import paths breaking (#469).
@@ -13,9 +13,9 @@ import { mkdir } from "node:fs/promises";
 import { chmodSync, statSync } from "node:fs";
 import { dirname } from "node:path";
 
-const assetGlob = new Bun.Glob("src/**/*.{md,xml,json,yml,yaml}");
+const assetGlob = new Bun.Glob("src/assets/**/*");
 for await (const src of assetGlob.scan(".")) {
-  const dest = src.replace(/^src\//, "dist/");
+  const dest = src.replace(/^src\/assets\//, "dist/assets/");
   await mkdir(dirname(dest), { recursive: true });
   await Bun.write(dest, Bun.file(src));
 }
