+++ src/assets/hints/cli-hints-full.md
