+++ src/assets/hints/cli-hints-short.md
