+++ src/assets/prompts/extract-session.md
