+++ src/assets/prompts/graph-extract-user-prompt.md
