+++ src/assets/tasks/graph-refresh-weekly.yml
