+++ src/assets/wiki/index-template.md
