+++ src/assets/wiki/schema-template.md
