+++ src/commands/extract-prompt.ts
@@ -16,8 +16,8 @@
  * the model emits get dropped before we parse.
  */
 
+import promptTemplate from "../assets/prompts/extract-session.md" with { type: "text" };
 import type { InlineRefMention, SessionData, SessionEvent } from "../integrations/session-logs/types";
-import promptTemplate from "../llm/prompts/extract-session.md" with { type: "text" };
 
 /**
  * JSON Schema for the structured extract output. Passed to `chatCompletion`
