+++ src/output/cli-hints.ts
@@ -12,7 +12,7 @@
  * `EMBEDDED_HINTS_FULL` (`--detail normal|full`, ~250 lines).
  */
 
-import EMBEDDED_HINTS_FULL from "./cli-hints-full.md" with { type: "text" };
-import EMBEDDED_HINTS from "./cli-hints-short.md" with { type: "text" };
+import EMBEDDED_HINTS_FULL from "../assets/hints/cli-hints-full.md" with { type: "text" };
+import EMBEDDED_HINTS from "../assets/hints/cli-hints-short.md" with { type: "text" };
 
 export { EMBEDDED_HINTS, EMBEDDED_HINTS_FULL };
