+++ src/wiki/wiki.ts
@@ -46,6 +46,7 @@
 import fs from "node:fs";
 import path from "node:path";
 import { parse as yamlParse } from "yaml";
+import ingestWorkflowTemplate from "../assets/wiki/ingest-workflow-template.md" with { type: "text" };
 import { akmSearch } from "../commands/search";
 import { isWithin, todayIso } from "../core/common";
 import { getSources, loadUserConfig, saveConfig } from "../core/config";
@@ -53,7 +54,6 @@ import { NotFoundError, UsageError } from "../core/errors";
 import { parseFrontmatter, parseFrontmatterBlock } from "../core/frontmatter";
 import { resolveSourceEntries, type SearchSource } from "../indexer/search-source";
 import type { SearchResponse, SourceSearchHit } from "../sources/types";
-import ingestWorkflowTemplate from "./ingest-workflow-template.md" with { type: "text" };
 import { buildIndexMd, buildLogMd, buildSchemaMd } from "./wiki-templates";
 
 // ── Constants ───────────────────────────────────────────────────────────────
