+++ src/workflows/authoring.ts
@@ -4,13 +4,13 @@
 
 import fs from "node:fs";
 import path from "node:path";
+import workflowTemplate from "../assets/workflows/workflow-template.md" with { type: "text" };
 import { resolveAssetPathFromName } from "../core/asset-spec";
 import { isWithin, resolveStashDir } from "../core/common";
 import { UsageError } from "../core/errors";
 import { warn } from "../core/warn";
 import { parseWorkflow } from "./parser";
 import type { WorkflowError } from "./schema";
-import workflowTemplate from "./workflow-template.md" with { type: "text" };
 
 const DEFAULT_WORKFLOW_TEMPLATE = renderWorkflowTemplate({
   title: "Example Workflow",
